requirements.txt
azure-functions
requests==2.23.0
adaptivecardbuilder==0.0.7